azdazda

+++ Vis/pages/🗓️_Fixtures.py
@@ -137,7 +137,7 @@ if selected_display == '📊Fixture Difficulty Rating':
         elif metric == "Average Goals Against (GA)":
             ga_matrix = ga.melt(id_vars='Team', var_name='GameWeek', value_name='GA')
             # Round GA values to 2 decimal places
-            ga_matrix['GA'] = ga_matrix['GA'].astype(int) 
+            ga_matrix['GA'] = ga_matrix['GA'].astype('float16') 
             filtered_ga_matrix = ga_matrix[(ga_matrix['GameWeek'] >= slider1) & (ga_matrix['GameWeek'] <= slider2)]
             pivot_ga_matrix = filtered_ga_matrix.pivot(index='Team', columns='GameWeek', values='GA')
             pivot_ga_matrix.columns = [f'GW {col}' for col in pivot_ga_matrix.columns].copy()
